i2c: stm32_i2c: correct standard mode min bus frequency
Prior this change a typo set the standard mode min bus frequency
to 8kHz whereas it should be 80kHz (80% of 100kHz).

This change explicitly sets the min/max bus frequency values
as 80% and 120% of the nominal frequency to overcome such issues.

This change also adds an explicit reference URL to the I2C
specification referred to from an inline comment.

Change-Id: Ie57652acede5cbb2181aa773af50ff3fabff73db
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Reviewed-on: https://gerrit.st.com/125416
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>
Tested-by: Yann GAUTIER <yann.gautier@st.com>