stm32mp1: allow debugger to reconstruct backtrace on panic
The first instruction of do_panic is to store the panic address in R6.
Filling this value in LR just before the endless loop in
plat_panic_handler allows a debugger to reconstruct the backtrace.
It is possible because no function uses R6 in the panic functions,
and the stack is not used.

Change-Id: I2398059f6be3e344a9a23c45edd834380646bf27
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/125218
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>