aarch32: add virtual timer control support
Define virtual timer resources for a platform to be able to control
virtual timers, even if TF-A does not enable them.
This is required for example when the secure world wants to reset
the system. Virtual timers enabled from non-secure world can
jeopardize the reset sequence due to pending interrupts.

Rename also CNTP_TVAL and CNTP_CTL registers to fit with
ARM Cortex-A7 Reference Manual.

Change-Id: I07ca8af84e83363495949098dc8baeff8893f7d9
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/128679
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>