i2c: stm32_i2c: compute dynamically I2C timings
Implement a dynamic calculation of I2C timings, depending on
device tree information (clock source, rise time, fall time).

Keep the capability to run the PMIC with a higher speed.

Release I2C driver implementation in dual license GPL-2.0+ OR
BSD-3-Clause as this driver is shared with packages as OP-TEE,
U-boot and the Linux Kernel.

Change-Id: Ia4fdcc678d857a8663f4583b7f8b2d342e39b86c
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Signed-off-by: Yann Gautier <yann.gautier@st.com>