stm32mp1: add ETZPC driver
ETZPC is for Extended TrustZone Protection Controller.
This is an extended version of the Arm TZPC [1].

The ETZPC configures TrustZone security in a SoC having bus masters
and slaves with programmable-security attributes (securable resources)
such as:
- On-chip RAM/ROM with programmable secure region size
- AHB and APB peripherals to be made secure
- AHB and APB peripherals to be set as isolated

As it differs a lot from the Arm TZPC, a dedicated driver is created.

[1] http://infocenter.arm.com/help/topic/com.arm.doc.dto0015a/index.html

Change-Id: Ia2caa49f09b9d2112fe5e9de47ee53bbfefedb52
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Signed-off-by: Lionel Debieve <lionel.debieve@st.com>
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>