stm32mp1: use device tree to determine board_id OTP
This OTP is filled only for STMicroelectronics boards, therefore the BSEC
device tree node for this should be moved to boards DTS files.
The function that displays this information is also reworked to use this
device tree information and macros.
A potential bug is also corrected to avoid non-null terminated string.

Change-Id: Ie4f268a4afff8cc3d63eb7ea31b4f44c8f488267
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/133034
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>