stm32mp1: tzc400: raise an FIQ in case of faulty access
TZC400 was configured to trig an exception when the non-secure world was
trying to access a secure memory. This leads to an AXI interrupt, which
doesn't show a lot of useful information.
Changing the TZC action in case of faulty access to TZC_ACTION_INT, now
triggers an FIQ. In this FIQ handler, more information is displayed, like
the address which caused the fault, and the faulty caller information
(secure/non-secure, privileged/unprivileged, read or write access).

Change-Id: I9a3c2e3797fbb8a462d3c35052a1ede624f1f70d
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/143267
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>