stm32mp1: Reduce MAX_XLAT_TABLES to 3
For STM32MP1, the address space is 4GB, which can be first divided
in 4 parts of 1GB. This LVL1 table is already mapped regardless
of MAX_XLAT_TABLES.
The SYSRAM needs fine grain MMU translation, by 4kB.
A LVL2 table is required to map the 2MB area containing the SYSRAM.
And a LVL3 table to map the 4kB pages of the SYSRAM.
The ROM code needs to be mapped for authentication, another 4kB table
is required.
For USB, an extra table is required to map MCU SRAM. An extra table is
also required in OP-TEE configuration. Both cases (USB and OP-TEE)
don't co-exist, so a logical OR can be implemented.

Change-Id: If7c1a06b989d0c3b3ee78dc6765e2631a89a4483
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/144662
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Tested-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>