stm32mp1: BL2 new clock initialization
If no pre-defined PLL1 settings in DT, or in case of standby wake up
(restore previously PLL1 settings from context in this case),
find the highest CPU frequency in the OPP table (in DT, compatible with
platform capabilities, or in structure restored in RAM), and set related
VDDCORE voltage (only if PMIC is present).
Then clock tree initialization can then be processed.

Change-Id: I15e50cc39ce784d29229e54bef79684fb2f4766c
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/149297
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>