stm32mp1: BL32 new clock initialization
On cold boot, compute and save all OPP PLL1 settings in SRAM.
The setting of the max OPP for CPU (voltage and frequency) was
previously done in BL2, but it is now moved to BL32. This is only done
if the board embeds a PMIC, and only for cold boot. After a Standby,
the PLL1 configuration is applied when restoring all clocks.

Change-Id: Idfad7b3a06b3eed62dfc61fa87f224fd440e506c
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>