bsec: handle OTP areas overlapping lower and upper regions
Use the device tree scanning loop to detect if an OTP, defined
with an index and a size in bytes, is overlapping lower and upper
regions of the OTP layout.
This shouldn't happen, but in case of, we consider only the upper
part, since access to lower part is always allowed.

Change-Id: I0f08e1dc9ac09d0abcfe224e41f2274975d7b026
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>