stm32mp1: set wakeup_standby in Standby with DDR in self-refresh case
The shutdown command can trigger a standby with DDR off, depending on the
DT configuration. In this case, the cold boot code must be done.
The wakeup_standby variable is then set to true if a return address is set
in the backup register.
This change avoids a panic in initialize_clock when pressing wake-up
button and the system was in STANDBY with DDR off.

Change-Id: I2b146340166bb077d8362818f9bcd8dedc23ee93
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/151752
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>