plat/stm32mp1: sp_min pm supports context V1 and V2
When TF-A BL2 boots an OP-TEE, the OP-TEE may be a old version that
provides a V1 PM context. In such case, precomputed PLL1 settings
are not available in the wakeup context and shall be recomputed.

With this stm32_get_pll1_settings_from_context() has a return value
that states if the PLL1 settings were restored.

Change-Id: Ia561e4ee42a788bc14a05e583650c5eb050131fd
Signed-off-by: Etienne Carriere <etienne.carriere@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/153628
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>