stm32mp1: low_power: manage self refresh in CSTOP for LPDDR2
Self Refresh (SR) is a significant power saving opportunity for SDRAM.
During SR, DLL's in DDRPHYC can be bypassed,and the DDRCTRL clocks can be
stopped.
Automatic SR (ASR) triggers on when there are no read/write for a
predefined time. This fits with the need in running mode.
Software SR (SSR) is controlled by software operation.

Switch to software self refresh when entering in CSTOP, and go back to
automatic self refresh when exiting from CSTOP.

Change-Id: I804872e934e13577233b41c2541374ca3d65041d
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/153878
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>