tools: stm32image: don't force base to strtol
Don't force the base of strtol, since it's able to select the base
automatically depending on the prefix of the value.
This does not breaks the current build script that extracts the
addresses, including the 0x prefix, from the map file.
This change helps using stm32image in shell scripts where the
addresses can be computed using the shell arithmetic expansion
"$((...))", that produces a value in base decimal.

Signed-off-by: Antonio Borneo <antonio.borneo@st.com>
Change-Id: I764c99be252ce868b8ef88a307907e9dba3fe76c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/154761
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>