stm32mp1: map DDR non-cacheable during its initialisation
This change avoids speculative accesses to DDR  whereas it is not reachable
and could lead to bus stall.
To achieve this the dynamic mapping in MMU is used.
A first mapping is done for DDR init with MT_NON_CACHEABLE attribute.
Once DDR access is setup, it is unmapped and a new mapping DDR is done with
cacheable attribute (through MT_MEMORY) to speed-up BL33 (or OP-TEE) load.

The disabling of cache during DDR tests is also removed, as now useless.

PLAT_XLAT_TABLES_DYNAMIC is activated globally as used in BL2 and BL32.

Change-Id: I22c48b4a48255ee264991c34ecbb15bfe87e67c3
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/154736
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>