pmic: stm32mp_pmic: only program voltage if not in DT range
When setting boot-on regulator, the voltage register is only re-programmed
if the read value in PMIC is not in the range provided in DT.
The dt_pmic_get_regulator_voltage() function is updated to read both
min and max voltage values.
The macro CMD_GET_VOLTAGE is renamed to CMD_GET_MIN_VOLTAGE to reflect
what it really does.

Change-Id: I8056e625e50ccf26ab32b2b84634bf53b840d9ac
Signed-off-by: Yann Gautier <yann.gautier@st.com>
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/159940
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>