stm32mp1: adaptations for STM32MP13 image header
The header must now include by default at least an extra padding header,
increasing the size of the header to 512 bytes (0x200).
This header will be placed at the end of SRAM3 by BootROM, letting the
whole SYSRAM to TF-A.
The boot context is now placed in SRAM2, hence this memory has to be
mapped in BL2 MMU.
The stm32image tool is updated to manage the extra padding header.
As this updated tool is also used to generate sp_min header, the
io_stm32image code has to be updated to manage the increased size of the
header in this case.

Change-Id: I50fcd43ecd0ba2076292b057566efe6809b9971a
Signed-off-by: Yann Gautier <yann.gautier@st.com>