stm32mp1: use FIP to load images
BL2 still uses the STM32 header binary format to be loaded from ROM code.
BL32 and BL33 and their respective device tree files are now put together
in a FIP file.
One DTB is created for each BL. To reduce their sizes, 2 new dtsi file are
in charge of removing useless nodes for a given BL. This is done because
BL2 and BL32 share the same device tree files base.

Change-Id: I1ffada0af58486d4cf6044511b51e56b52269817
Signed-off-by: Yann Gautier <yann.gautier@st.com>