tools: allow to override OPENSSL_DIR for external SDK
In case of external SDK management, we must be able to specify
a library path for OPENSSL libraries. By removing the default
path set to /usr, we can select a OPENSSL_DIR path from
the command line.
Add the libpthread dependency because openssl is linked to this library.
The libpthread includes some specific options that needs to be aligned
to the openssl library used.

Signed-off-by: Lionel Debieve <lionel.debieve@st.com>
Change-Id: Ic81b2e7a827e3ce6794b40aa6c902adf076b977c
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/189679
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@st.com>