stm32mp1: move DDR and cache management in low-power critical section
The actions on DDR self-refresh and cache flush should be done in critic
power functions, that always run from embedded SYSRAM.
The DDR area that will be used for training is also saved in this critic
power part. As cstop_critic_enter() should now be executed when going to
Standby, change the first parameter when calling stm32_pwr_down_wfi() to
true in stm32_system_off() function.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: Ia5eb4e7b0e1030c7936aa9f742b99d8c0222c98a
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/206181
Reviewed-by: MPUOSTL
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>
Reviewed-by: Pascal PAILLET <p.paillet@st.com>