stm32mp1: use a fixed address for monitor in DDR
DDR_SEC_SIZE is define in the platform. It will be used in both fw-config
device tree file, and in the code to populate reserved memory in non-secure
device tree.
The part of code in charge of populating non-secure DT is put under
STM32MP_SP_MIN_IN_DDR flag and simplified.

Change-Id: I7ec38eae2f10585a9a3853ec6a9b34190fd1467d
Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/206415
Reviewed-by: MPUOSTL
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>