plat/st: improve image loading from MMC with FIP
Instead of using a scratch buffer of 512 bytes, we can directly use the
image address and max size. The mmc_block_dev_spec struct info is then
overwritten for each image with this info, except FW_CONFIG and GPT table
which will still use the scratch buffer.
This allows using multiple blocks read on MMC, and so improves the boot
time.
A cache invalidate is required for the remaining data not used from the
last block read. Except for FW_CONFIG_ID, it is not required, as it is
in scratch buffer in SYSRAM, and also because bl_mem_params struct is
overwritten in this case.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: If3ecfdfe35bb9db66284036ca49c4bd1be4fd121
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/211727
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Nicolas LE BAYON <nicolas.le.bayon@st.com>