clk: stm32mp1_clk: skip always-on clocks in stm32mp1_clock_resume()
Gateable clocks, listed in stm32mp1_clk_gate[] table, get resumed in
stm32mp1_clock_resume() function, when returning from Standby state.
If a clock is always-on, like RTC, it doesn't get a refcount. So the
resume function tries to disable it. Such clocks should then be skipped
in the gate resume loop.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: Iac7e977fa4916614522a2270a44f08f243238943
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/230363
Reviewed-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: CIBUILD <MDG-smet-aci-builds@list.st.com>
Reviewed-by: Lionel DEBIEVE <lionel.debieve@foss.st.com>